Switch to using ML-IAP for LAMMPS, at least on Polaris (#215)
* Change the build instructions for Polaris

* Add cuEQ to environment, use when compiling model

* Support MACE ML-IAP in the MACE runner

* Document ML-IAP runner

* Explain how to configure a MACE model

* Update the component test

+++ input-files/mace/get-macemp-0a.sh
@@ -1,4 +1,6 @@
 #! /bin/bash
 
-wget https://github.com/ACEsuit/mace-mp/releases/download/mace_mp_0/2023-12-03-mace-128-L1_epoch-199.model -O mace-mp0_medium
-mace_create_lammps_model mace-mp0_medium
+if [ ! -e mace-mp0_medium ]; then
+  wget https://github.com/ACEsuit/mace-mp/releases/download/mace_mp_0/2023-12-03-mace-128-L1_epoch-199.model -O mace-mp0_medium
+fi
+mace_create_lammps_model mace-mp0_medium --format mliap --dtype float32
